move to a new sham lib
