+++ src/shamdag/include/shamdag/graph2dot.hpp
@@ -1,0 +1,24 @@
+// -------------------------------------------------------//
+//
+// SHAMROCK code for hydrodynamics
+// Copyright (c) 2021-2024 Timothée David--Cléris <tim.shamrock@proton.me>
+// SPDX-License-Identifier: CeCILL Free Software License Agreement v2.1
+// Shamrock is licensed under the CeCILL 2.1 License, see LICENSE for more information
+//
+// -------------------------------------------------------//
+
+#pragma once
+
+/**
+ * @file graph2dot.hpp
+ * @author Timothée David--Cléris (tim.shamrock@proton.me)
+ * @brief
+ *
+ */
+
+#include "shamdag/shamdag.hpp"
+#include <memory>
+#include <string>
+#include <vector>
+
+std::string get_node_dot_graph(std::vector<std::shared_ptr<INode>> endpoints);
